covert to int

+++ experiments/mnist.py
@@ -207,7 +207,8 @@ class MNIST():
         return x.astype(int)
 
     def distance(self, y):
-        print(int(y))
+        y = y.astype(int)
+        print(y)
         print(self.y_train)
         sys.exit()
         return 1/self.y_train.shape[0] * sum(np.invert(np.logical_xor(self.y_train, y)))
